backend/requirements.txt
# Web 框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 图像处理
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0

# 工具
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.0

# SAM 模型 (按需安装)
# torch>=2.2.0
# torchvision>=0.17.0
# segment-anything  (SAM1)
# sam2              (SAM2)
# sam-hq            (SAM-HQ)
# sam3              (SAM3)
